+++ packages/backend/src/server/api/endpoints/i/notifications.ts
@@ -3,7 +3,7 @@ import { ID } from '@/misc/cafy-id';
 import { readNotification } from '../../common/read-notification';
 import define from '../../define';
 import { makePaginationQuery } from '../../common/make-pagination-query';
-import { generateMutedInstanceQuery } from '../../common/generate-muted-instance-query';
+import { generateMutedInstanceNotificationQuery } from '../../common/generate-muted-instance-query';
 import { Notifications, Followings, Mutings, Users } from '@/models/index';
 import { notificationTypes } from '@/types';
 import read from '@/services/note/read';
@@ -102,7 +102,7 @@ export default define(meta, async (ps, user) => {
 	}));
 	query.setParameters(mutingQuery.getParameters());
 
-	// generateMutedInstanceQuery(query, user);
+	generateMutedInstanceNotificationQuery(query, user);
 
 	query.andWhere(new Brackets(qb => { qb
 		.where(`notification.notifierId NOT IN (${ suspendedQuery.getQuery() })`)