+++ CHANGELOG.md
@@ -229,6 +229,9 @@
 
 ## 12.89.1 (2021/08/24)
 
+### Features
+- Added a user-level instance mute in user settings
+
 ### Improvements
 - クライアントのデザインの調整